+++ Modules/Ruleset.ProgramInfo/Public/Get-SystemApp.ps1
@@ -146,7 +146,7 @@ function Get-SystemApp
 		Select-Object -ExpandProperty SystemDrive
 
 		$SystemDrive = $SystemDrive.TrimEnd(":")
-		$DomainPath = "\\$Domain\$SystemDrive`$\"
+		$DomainPath = "\\$Domain\$SystemDrive$"
 	}
 
 	if (!$Apps)