+++ Modules/Ruleset.ProgramInfo/Public/Get-UserSoftware.ps1
@@ -39,8 +39,14 @@ User name in form of "USERNAME"
 .PARAMETER Domain
 NETBIOS Computer name in form of "COMPUTERNAME"
 
+.PARAMETER Credential
+Specifies the credential object to use for authentication
+
 .PARAMETER CimSession
 Specifies the CIM session to use
+
+.PARAMETER Session
+Specifies the PS session to use
 
 .EXAMPLE
 PS> Get-UserSoftware "User"
@@ -71,22 +77,46 @@ function Get-UserSoftware
 		[Alias("ComputerName", "CN")]
 		[string] $Domain = [System.Environment]::MachineName,
 
-		[Parameter(ParameterSetName = "CimSession")]
-		[CimSession] $CimSession
+		[Parameter(ParameterSetName = "Session")]
+		[CimSession] $CimSession,
+
+		[Parameter(ParameterSetName = "Session")]
+		[System.Management.Automation.Runspaces.PSSession] $Session
 	)
 
 	Write-Debug -Message "[$($MyInvocation.InvocationName)] Caller = $((Get-PSCallStack)[1].Command) ParameterSet = $($PSCmdlet.ParameterSetName):$($PSBoundParameters | Out-String)"
 
 	[hashtable] $CimParams = @{}
-	if ($PSCmdlet.ParameterSetName -eq "CimSession")
+	[hashtable] $SessionParams = @{}
+
+	if ($PSCmdlet.ParameterSetName -eq "Session")
 	{
+		if ($Session.ComputerName -ne $CimSession.ComputerName)
+		{
+			Write-Error -Category InvalidArgument -TargetObject $CimSession `
+				-Message "Session and CimSession must be targeting same computer"
+			return
+		}
+
 		$Domain = $CimSession.ComputerName
 		$CimParams.CimSession = $CimSession
+		$SessionParams.Session = $Session
 	}
 	else
 	{
 		$Domain = Format-ComputerName $Domain
-		$CimParams.ComputerName = $Domain
+
+		# Avoiding NETBIOS ComputerName for localhost means no need for WinRM to listen on HTTP
+		if ($Domain -ne [System.Environment]::MachineName)
+		{
+			$CimParams.ComputerName = $Domain
+			$SessionParams.ComputerName = $Domain
+
+			if ($Credential)
+			{
+				$SessionParams.Credential = $Credential
+			}
+		}
 	}
 
 	if (Test-Computer $Domain)
@@ -94,34 +124,90 @@ function Get-UserSoftware
 		$Principal = Get-PrincipalSID $User @CimParams
 		if (!$Principal) { return }
 
-		# TODO: Need to load key for user which is not logged on
-		$HKU = $Principal.SID
-		$HKU += "\Software\Microsoft\Windows\CurrentVersion\Uninstall"
+		$UserSID = $Principal.SID
 		$RegistryHive = [Microsoft.Win32.RegistryHive]::Users
 
 		try
 		{
 			Write-Verbose -Message "[$($MyInvocation.InvocationName)] Accessing registry on computer: $Domain"
 			$RemoteKey = [Microsoft.Win32.RegistryKey]::OpenRemoteBaseKey($RegistryHive, $Domain)
+		}
+		catch
+		{
+			Write-Error -ErrorRecord $_
+			return
+		}
 
-			Write-Verbose -Message "[$($MyInvocation.InvocationName)] Opening root key HKU:$HKU"
+		# Check if target user is logged in (user reg hive is loaded)
+		$TempKey = $null
+
+		if ([array]::Find($RemoteKey.GetSubKeyNames(), [System.Predicate[string]] { $UserSID -eq $args[0] }))
+		{
+			$HKU = "$UserSID\Software\Microsoft\Windows\CurrentVersion\Uninstall"
+		}
+		else
+		{
+			Write-Verbose -Message "[$($MyInvocation.InvocationName)] User '$User' is not logged into system"
+
+			[string] $SystemDrive = Get-CimInstance -Class Win32_OperatingSystem @CimParams |
+			Select-Object -ExpandProperty SystemDrive
+			$UserRegConfig = "$SystemDrive\Users\$User\NTUSER.DAT"
+
+			if ($Domain -eq [System.Environment]::MachineName)
+			{
+				$PathToTest = $UserRegConfig
+			}
+			else
+			{
+				$PathToTest = "\\$Domain\$($SystemDrive.TrimEnd(":"))$\Users\$User\NTUSER.DAT"
+			}
+
+			# NOTE: Using User-UserName instead of SID to minimize the chance of existing key with same name
+			$TempKey = "User-$User"
+
+			if (Test-Path -Path $PathToTest)
+			{
+				Write-Verbose -Message "[$($MyInvocation.InvocationName)] Loading offline hive for user '$User' to HKU\$TempKey"
+
+				# NOTE: Invoke-Process is needed to make the command finish it's job and print status
+				$Status = Invoke-Process -NoNewWindow reg.exe -ArgumentList "load HKU\$TempKey $UserRegConfig" -Raw @SessionParams
+
+				Write-Debug -Message "[$($MyInvocation.InvocationName)] reg load status is '$Status'"
+				$HKU = "$TempKey\Software\Microsoft\Windows\CurrentVersion\Uninstall"
+			}
+			else
+			{
+				$RemoteKey.Dispose()
+				Write-Warning -Message "[$($MyInvocation.InvocationName)] Failed to locate user registry config '$UserRegConfig'"
+				return
+			}
+		}
+
+		try
+		{
+			Write-Verbose -Message "[$($MyInvocation.InvocationName)] Opening root key HKU:\$HKU"
 			$RootKey = $RemoteKey.OpenSubkey($HKU, $RegistryPermission, $RegistryRights)
 
 			if (!$RootKey)
 			{
-				Write-Warning -Message "[$($MyInvocation.InvocationName)] Following registry key does not exist: $HKU"
-				$RemoteKey.Dispose()
-				return
+				throw [System.Data.ObjectNotFoundException]::new("Following registry key does not exist: $HKU")
 			}
 		}
 		catch
 		{
-			if ($RemoteKey)
+			Write-Debug -Message "[$($MyInvocation.InvocationName)] Exception opening registry root key: HKU:\$HKU"
+
+			if ($TempKey)
 			{
-				$RemoteKey.Dispose()
+				Write-Verbose -Message "[$($MyInvocation.InvocationName)] Unload and release hive HKU:$TempKey"
+
+				[gc]::collect()
+				$Status = Invoke-Process reg.exe -NoNewWindow -ArgumentList "unload HKU\$TempKey" -Raw @SessionParams
+				Write-Debug -Message "[$($MyInvocation.InvocationName)] reg unload status is '$Status'"
 			}
 
-			Write-Error -ErrorRecord $_
+			Write-Warning -Message "[$($MyInvocation.InvocationName)] $($_.Exception.Message)"
+			$RemoteKey.Dispose()
 			return
 		}
 
@@ -141,6 +227,10 @@ function Get-UserSoftware
 			if ([string]::IsNullOrEmpty($InstallLocation))
 			{
 				Write-Warning -Message "[$($MyInvocation.InvocationName)] Failed to read registry entry $HKUSubKey\InstallLocation"
+
+				# NOTE: each key accessed after 'reg load' has to be closed to release handle, if not 'reg unload' fails with "Access is denied"
+				# TODO: Other functions in ProgramInfo module should implement closing keys to release handles.
+				$SubKey.Close()
 				continue
 			}
 
@@ -160,6 +250,20 @@ function Get-UserSoftware
 				RegistryKey = "HKU:\$HKU"
 				PSTypeName = "Ruleset.ProgramInfo"
 			}
+
+			$SubKey.Close()
+		}
+
+		$RootKey.Close()
+
+		if ($TempKey)
+		{
+			Write-Verbose -Message "[$($MyInvocation.InvocationName)] Unload and release hive HKU\$TempKey"
+
+			# NOTE: This is not strictly needed but is recommended before calling reg unload
+			[gc]::collect()
+			$Status = Invoke-Process reg.exe -NoNewWindow -ArgumentList "unload HKU\$TempKey" -Raw @SessionParams
+			Write-Debug -Message "[$($MyInvocation.InvocationName)] reg unload status is '$Status'"
 		}
 
 		$RemoteKey.Dispose()